Fix compilation warning with newer gcc

+++ src/rdkafka_conf.h
@@ -64,7 +64,7 @@ rd_kafka_compression2str (rd_kafka_compression_t compr) {
         };
         static RD_TLS char ret[32];
 
-        if (compr < 0 || compr >= RD_KAFKA_COMPRESSION_NUM) {
+        if ((int)compr < 0 || compr >= RD_KAFKA_COMPRESSION_NUM) {
                 rd_snprintf(ret, sizeof(ret),
                             "codec0x%x?", (int)compr);
                 return ret;
